+++ src/aim/models/loader.py
@@ -8,7 +8,7 @@ from aim.models.metrics import MonitorConfig, Metric, ec2core, CloudWatchAlarm, 
 from aim.models.metrics import LogSets, LogSet, LogCategory, LogSource, CWAgentLogSource
 from aim.models.networks import NetworkEnvironment, Environment, EnvironmentDefault, \
     EnvironmentRegion, Segment, Network, VPC, NATGateway, VPNGateway, PrivateHostedZone, \
-    SecurityGroup, IngressRule, EgressRule
+    SecurityGroup, IngressRule, EgressRule, Route53, Route53HostedZone
 from aim.models.project import Project, Credentials
 from aim.models.apps import ApplicationEngines, ApplicationEngine, Application, ResourceGroup
 from aim.models.resources import CodePipeBuildDeploy, ASG, Resource, Resources,LBApplication, \
@@ -17,7 +17,7 @@ from aim.models.resources import CodePipeBuildDeploy, ASG, Resource, Resources,L
 from aim.models.governance import Governance, GovernanceService, GovernanceMonitoring
 from aim.models.iam import IAMs, IAM, ManagedPolicy, Role, Policy, AssumeRolePolicy, Statement
 from aim.models.base import get_all_fields
-from aim.models.storages import RDS
+from aim.models.storages import RDS, CodeCommit, CodeCommitRepository
 from aim.models.accounts import Account, AdminIAMUser
 from aim.models.references import TextReference
 from aim.models.vocabulary import aws_regions
@@ -47,7 +47,8 @@ RESOURCES_CLASS_MAP = {
     'LBApplication': LBApplication,
     'EC2': EC2,
     'Lambda': Lambda,
-    'ManagedPolicy': ManagedPolicy
+    'ManagedPolicy': ManagedPolicy,
+    'S3Bucket': S3Bucket
 }
 
 SUB_TYPES_CLASS_MAP = {
@@ -129,6 +130,9 @@ SUB_TYPES_CLASS_MAP = {
     # Accounts
     Account: {
         'admin_iam_users': ('named_dict', AdminIAMUser)
+    },
+    Route53: {
+        'hosted_zones': ('named_dict', Route53HostedZone)
     }
 }
 
@@ -389,7 +393,8 @@ class ModelLoader():
             "MonitorConfig": self.instantiate_monitor_config,
             "Accounts": self.instantiate_accounts,
             "NetworkEnvironments": self.instantiate_network_environments,
-            "Governance": self.instantiate_governance
+            "Governance": self.instantiate_governance,
+            "Services": self.instantiate_services
         }
         self.yaml = YAML(typ="safe", pure=True)
         self.yaml.default_flow_sytle = False
@@ -710,6 +715,35 @@ class ModelLoader():
         # Under Construction
         #for [service_id, service_config] in config['services'].items():
         #    self.load_governance_service(service_id)
+
+    def instantiate_route53(self, config):
+        obj = Route53(config)
+        #breakpoint()
+        apply_attributes_from_config(obj, config)
+        return obj
+
+    def instantiate_codecommit(self, config):
+        codecommit_obj = CodeCommit()
+        codecommit_obj.repository_groups = {}
+        for group_id in config.keys():
+            group_config = config[group_id]
+            codecommit_obj.repository_groups[group_id] = {}
+            for repo_id in group_config.keys():
+                repo_config = group_config[repo_id]
+                repo_obj = CodeCommitRepository(repo_config['name'], codecommit_obj)
+                apply_attributes_from_config(repo_obj, repo_config)
+                codecommit_obj.repository_groups[group_id][repo_id] = repo_obj
+
+        codecommit_obj.gen_repo_by_account()
+        return codecommit_obj
+
+
+    def instantiate_services(self, name, config):
+        if name == "Route53":
+            self.project['route53'] = self.instantiate_route53(config)
+        elif name == "CodeCommit":
+            self.project['codecommit'] = self.instantiate_codecommit(config)
+        return
 
     def instantiate_monitor_config(self, name, config):
         """